+++ ogcore/SS.py
@@ -39,7 +39,7 @@ def euler_equation_solver(guesses, *args):
 
     Args:
         guesses (Numpy array): initial guesses for b and n, length 2S
-        args (tuple): tuple of arguments (r, w, p_tilde, bq, TR, factor, j, p)
+        args (tuple): tuple of arguments (r, w, p_tilde, p_i, bq, TR, factor, j, p)
         r (scalar): real interest rate
         w (scalar): real wage rate
         p_tilde (scalar): composite good price
@@ -54,7 +54,7 @@ def euler_equation_solver(guesses, *args):
         errros (Numpy array): errors from FOCs, length 2S
 
     """
-    (r, w, p_tilde, bq, rm, tr, ubi, factor, j, p) = args
+    r, w, p_tilde, p_i, bq, rm, tr, ubi, factor, j, p = args
 
     b_guess = np.array(guesses[: p.S])
     n_guess = np.array(guesses[p.S :])
@@ -67,6 +67,7 @@ def euler_equation_solver(guesses, *args):
         r,
         w,
         p_tilde,
+        p_i,
         b_s,
         b_splus1,
         n_guess,
@@ -88,6 +89,7 @@ def euler_equation_solver(guesses, *args):
         r,
         w,
         p_tilde,
+        p_i,
         b_s,
         b_splus1,
         n_guess,
@@ -143,6 +145,7 @@ def euler_equation_solver(guesses, *args):
         r,
         w,
         p_tilde,
+        p_i,
         b_s,
         b_splus1,
         n_guess,
@@ -150,6 +153,7 @@ def euler_equation_solver(guesses, *args):
         rm,
         taxes,
         p.e[-1, :, j],
+        p.tau_c[-1, :],
         p,
     )
     mask6 = cons < 0
@@ -164,6 +168,7 @@ def solve_for_j(
     r_p,
     w,
     p_tilde,
+    p_i,
     bq_j,
     rm_j,
     tr_j,
@@ -180,6 +185,7 @@ def solve_for_j(
         r_p (scalar): return on household investment portfolio
         w (scalar): real wage rate
         p_tilde (scalar): composite good price
+        p_i (Numpy array): prices for consumption good i
         bq_j (Numpy array): bequest amounts by age, length S
         rm_j (Numpy array): remittance amounts by age, length S
         tr_j (Numpy array): government transfer amount by age, length S
@@ -200,6 +206,7 @@ def solve_for_j(
             r_p,
             w,
             p_tilde,
+            p_i,
             bq_j,
             rm_j,
             tr_j,
@@ -314,6 +321,7 @@ def inner_loop(outer_loop_vars, p, client):
                 r_p,
                 w,
                 p_tilde,
+                p_i,
                 bq[:, j],
                 rm[:, j],
                 tr[:, j],
@@ -346,6 +354,7 @@ def inner_loop(outer_loop_vars, p, client):
                     r_p,
                     w,
                     p_tilde,
+                    p_i,
                     bq[:, j],
                     rm[:, j],
                     tr[:, j],
@@ -365,6 +374,7 @@ def inner_loop(outer_loop_vars, p, client):
                 r_p,
                 w,
                 p_tilde,
+                p_i,
                 bq[:, j],
                 rm[:, j],
                 tr[:, j],
@@ -408,7 +418,7 @@ def inner_loop(outer_loop_vars, p, client):
         None,
         False,
         "SS",
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         etr_params_3D,
         p,
     )
@@ -416,21 +426,25 @@ def inner_loop(outer_loop_vars, p, client):
         r_p,
         w,
         p_tilde,
+        p_i,
         b_s,
         b_splus1,
         nssmat,
         bq,
         rm,
         net_tax,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
+        p.tau_c[-1, :],
         p,
     )
-    c_i = household.get_ci(c_s, p_i, p_tilde, p.tau_c[-1, :], p.alpha_c)
+    c_i = household.get_ci(
+        c_s, p_i, p_tilde, p.tau_c[-1, :], p.alpha_c, p.c_min
+    )
     L = aggr.get_L(nssmat, p, "SS")
     B = aggr.get_B(bssmat, p, "SS", False)
 
     # Find gov't debt
-    r_gov = fiscal.get_r_gov(r, p, "SS")
+    r_gov = fiscal.get_r_gov(r, p.debt_ratio_ss, p, "scalar", t=-1)
     D, D_d, D_f, new_borrowing, _, new_borrowing_f = fiscal.get_D_ss(
         r_gov, Y, p
     )
@@ -481,7 +495,7 @@ def inner_loop(outer_loop_vars, p, client):
         new_r = firm.get_r(Y_vec[-1], K_vec[-1], p_m, p, "SS", -1)
     new_w = firm.get_w(Y_vec[-1], L_vec[-1], p_m, p, "SS")
 
-    new_r_gov = fiscal.get_r_gov(new_r, p, "SS")
+    new_r_gov = fiscal.get_r_gov(new_r, p.debt_ratio_ss, p, "scalar", t=-1)
     # now get accurate measure of debt service cost
     (
         D,
@@ -543,7 +557,7 @@ def inner_loop(outer_loop_vars, p, client):
         None,
         False,
         "SS",
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         etr_params_3D,
         p,
     )
@@ -551,13 +565,15 @@ def inner_loop(outer_loop_vars, p, client):
         new_r_p,
         new_w,
         new_p_tilde,
+        new_p_i,
         b_s,
         bssmat,
         nssmat,
         new_bq,
         new_rm,
         taxss,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
+        p.tau_c[-1, :],
         p,
     )
     (
@@ -586,7 +602,7 @@ def inner_loop(outer_loop_vars, p, client):
         ubi,
         theta,
         etr_params_3D,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         p,
         None,
         "SS",
@@ -813,7 +829,7 @@ def SS_solver(
     K_vec_ss = new_K_vec
     L_vec_ss = new_L_vec
     Y_vec_ss = new_Y_vec
-    r_gov_ss = fiscal.get_r_gov(rss, p, "SS")
+    r_gov_ss = fiscal.get_r_gov(rss, p.debt_ratio_ss, p, "scalar", t=-1)
     p_m_ss = new_p_m
     p_i_ss = np.dot(p.io_matrix, p_m_ss)
     p_tilde_ss = aggr.get_ptilde(p_i_ss, p.tau_c[-1, :], p.alpha_c)
@@ -898,6 +914,10 @@ def SS_solver(
         np.reshape(p.labor_income_tax_noncompliance_rate[-1, :], (1, p.J)),
         (p.S, 1),
     )
+    income_tax_filer_2D = np.tile(
+        np.reshape(p.income_tax_filer[-1, :], (1, p.J)),
+        (p.S, 1),
+    )
     mtry_ss = tax.MTR_income(
         r_p_ss,
         wss,
@@ -905,10 +925,11 @@ def SS_solver(
         nssmat,
         factor,
         True,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         etr_params_3D,
         mtry_params_3D,
         capital_noncompliance_rate_2D,
+        income_tax_filer_2D,
         p,
     )
     mtrx_ss = tax.MTR_income(
@@ -918,10 +939,11 @@ def SS_solver(
         nssmat,
         factor,
         False,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         etr_params_3D,
         mtrx_params_3D,
         labor_noncompliance_rate_2D,
+        income_tax_filer_2D,
         p,
     )
     etr_ss = tax.ETR_income(
@@ -930,10 +952,11 @@ def SS_solver(
         bssmat_s,
         nssmat,
         factor,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         etr_params_3D,
         labor_noncompliance_rate_2D,
         capital_noncompliance_rate_2D,
+        income_tax_filer_2D,
         p,
     )
 
@@ -951,7 +974,7 @@ def SS_solver(
         None,
         False,
         "SS",
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         etr_params_3D,
         p,
     )
@@ -990,13 +1013,15 @@ def SS_solver(
         r_p_ss,
         wss,
         p_tilde_ss,
+        p_i_ss,
         bssmat_s,
         bssmat_splus1,
         nssmat,
         bqssmat,
         rmssmat,
         taxss,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
+        p.tau_c[-1, :],
         p,
     )
     c_i = household.get_ci(
@@ -1005,6 +1030,7 @@ def SS_solver(
         p_tilde_ss,
         p.tau_c[-1, :],
         p.alpha_c,
+        p.c_min,
         "SS",
     )
     sales_tax_ss = tax.cons_tax_liab(c_i, p_i_ss, p, "SS")
@@ -1013,7 +1039,7 @@ def SS_solver(
     )
     Css = aggr.get_C(cssmat, p, "SS")
     c_i_ss_mat = household.get_ci(
-        cssmat, p_i_ss, p_tilde_ss, p.tau_c[-1, :], p.alpha_c
+        cssmat, p_i_ss, p_tilde_ss, p.tau_c[-1, :], p.alpha_c, p.c_min
     )
     C_vec_ss = np.zeros(p.I)
     for i_ind in range(
@@ -1051,7 +1077,7 @@ def SS_solver(
         ubissmat,
         theta,
         etr_params_3D,
-        np.squeeze(p.e[-1, :, :]),
+        np.squeeze(p.e[-1, :, :]).reshape((p.S, p.J)),
         p,
         None,
         "SS",
@@ -1235,7 +1261,7 @@ def SS_fsolve(guesses, *args):
             implied outer loop variables
 
     """
-    (bssmat, nssmat, TR_ss, Ig_baseline, factor_ss, p, client) = args
+    bssmat, nssmat, TR_ss, Ig_baseline, factor_ss, p, client = args
 
     # Rename the inputs
     r_p = guesses[0]