+++ ogcore/household.py
@@ -363,7 +363,22 @@ def get_rm(RM, j, p, method):
     return rm
 
 
-def get_cons(r_p, w, p_tilde, b, b_splus1, n, bq, rm, net_tax, e, p):
+def get_cons(
+    r_p,
+    w,
+    p_tilde,
+    p_i,
+    b,
+    b_splus1,
+    n,
+    bq,
+    rm,
+    net_tax,
+    e,
+    tau_c,
+    p,
+    method="SS",
+):
     r"""
     Calculate household composite consumption.
 
@@ -381,6 +396,7 @@ def get_cons(r_p, w, p_tilde, b, b_splus1, n, bq, rm, net_tax, e, p):
         r_p (array_like): the real interest rate
         w (array_like): the real wage rate
         p_tilde (array_like): the ratio of real GDP to nominal GDP
+        p_i (Numpy array): prices for consumption good i
         b (Numpy array): household savings
         b_splus1 (Numpy array): household savings one period ahead
         n (Numpy array): household labor supply
@@ -388,12 +404,32 @@ def get_cons(r_p, w, p_tilde, b, b_splus1, n, bq, rm, net_tax, e, p):
         rm (Numpy array): household remittances received
         net_tax (Numpy array): household net taxes paid
         e (Numpy array): effective labor units
+        tau_c (Numpy array): consumption tax rates
         p (OG-Core Specifications object): model parameters
+        method (str): adjusts calculation dimensions based on 'SS' or 'TPI'
 
     Returns:
         cons (Numpy array): household consumption
 
     """
+    if method == "SS":
+        min_cons_expenditure = np.sum((1 + tau_c) * p_i * p.c_min)
+    elif method == "TPI_scalar":
+        min_cons_expenditure = np.sum((1 + tau_c) * p_i * p.c_min)
+    else:  # TPI case
+        min_cons_expenditure = np.sum(
+            (1 + tau_c) * p_i * p.c_min.reshape((1, p.I)), axis=1
+        )
+        length = r_p.shape[0]
+        # reshape if needed for broadcasting
+        n_dims = np.ndim(bq)
+        if n_dims == 1:
+            min_cons_expenditure = min_cons_expenditure.reshape((length,))
+        elif n_dims == 2:
+            min_cons_expenditure = min_cons_expenditure.reshape((length, 1))
+        else:
+            min_cons_expenditure = min_cons_expenditure.reshape((length, 1, 1))
+
     cons = (
         (1 + r_p) * b
         + w * e * n
@@ -401,12 +437,13 @@ def get_cons(r_p, w, p_tilde, b, b_splus1, n, bq, rm, net_tax, e, p):
         + rm
         - net_tax
         - b_splus1 * np.exp(p.g_y)
+        - min_cons_expenditure
     ) / p_tilde
 
     return cons
 
 
-def get_ci(c_s, p_i, p_tilde, tau_c, alpha_c, method="SS"):
+def get_ci(c_s, p_i, p_tilde, tau_c, alpha_c, c_min, method="SS"):
     r"""
     Compute consumption of good i given amount of composite consumption
     and prices.
@@ -420,6 +457,7 @@ def get_ci(c_s, p_i, p_tilde, tau_c, alpha_c, method="SS"):
         p_tilde (array_like): composite good price
         tau_c (array_like): consumption tax rate
         alpha_c (array_like): consumption share parameters
+        c_min (array_like): minimum consumption levels for each good i
         method (str): adjusts calculation dimensions based on 'SS' or 'TPI'
 
     Returns:
@@ -431,10 +469,11 @@ def get_ci(c_s, p_i, p_tilde, tau_c, alpha_c, method="SS"):
         J = c_s.shape[1]
         tau_c = tau_c.reshape(I, 1, 1)
         alpha_c = alpha_c.reshape(I, 1, 1)
+        c_min = c_min.reshape(I, 1, 1)
         p_tilde.reshape(1, 1, 1)
         p_i = p_i.reshape(I, 1, 1)
         c_s = c_s.reshape(1, S, J)
-        c_si = alpha_c * (((1 + tau_c) * p_i) / p_tilde) ** (-1) * c_s
+        c_si = alpha_c * (((1 + tau_c) * p_i) / p_tilde) ** (-1) * c_s + c_min
     else:  # Time path case
         I = alpha_c.shape[0]
         T = p_i.shape[0]
@@ -442,10 +481,11 @@ def get_ci(c_s, p_i, p_tilde, tau_c, alpha_c, method="SS"):
         J = c_s.shape[2]
         tau_c = tau_c.reshape(T, I, 1, 1)
         alpha_c = alpha_c.reshape(1, I, 1, 1)
+        c_min = c_min.reshape(1, I, 1, 1)
         p_tilde = p_tilde.reshape(T, 1, 1, 1)
         p_i = p_i.reshape(T, I, 1, 1)
         c_s = c_s.reshape(T, 1, S, J)
-        c_si = alpha_c * (((1 + tau_c) * p_i) / p_tilde) ** (-1) * c_s
+        c_si = alpha_c * (((1 + tau_c) * p_i) / p_tilde) ** (-1) * c_s + c_min
     return c_si
 
 
@@ -453,6 +493,7 @@ def FOC_savings(
     r,
     w,
     p_tilde,
+    p_i,
     b,
     b_splus1,
     n,
@@ -486,6 +527,7 @@ def FOC_savings(
         r (array_like): the real interest rate
         w (array_like): the real wage rate
         p_tilde (array_like): composite good price
+        p_i (Numpy array): prices for consumption good i
         b (Numpy array): household savings
         b_splus1 (Numpy array): household savings one period ahead
         b_splus2 (Numpy array): household savings two periods ahead
@@ -517,15 +559,21 @@ def FOC_savings(
         beta = p.beta[j]
         if method == "SS":
             tax_noncompliance = p.capital_income_tax_noncompliance_rate[-1, j]
+            income_tax_filer = p.income_tax_filer[-1, j]
+            wealth_tax_filer = p.wealth_tax_filer[-1, j]
             e = np.squeeze(p.e[-1, :, j])
         elif method == "TPI_scalar":
             tax_noncompliance = p.capital_income_tax_noncompliance_rate[0, j]
+            income_tax_filer = p.income_tax_filer[0, j]
+            wealth_tax_filer = p.wealth_tax_filer[0, j]
             e = np.squeeze(p.e[0, :, j])
         else:
             length = r.shape[0]
             tax_noncompliance = p.capital_income_tax_noncompliance_rate[
                 t : t + length, j
             ]
+            income_tax_filer = p.income_tax_filer[t : t + length, j]
+            wealth_tax_filer = p.wealth_tax_filer[t : t + length, j]
             e_long = np.concatenate(
                 (
                     p.e,
@@ -539,15 +587,21 @@ def FOC_savings(
         beta = p.beta
         if method == "SS":
             tax_noncompliance = p.capital_income_tax_noncompliance_rate[-1, :]
+            income_tax_filer = p.income_tax_filer[-1, :]
+            wealth_tax_filer = p.wealth_tax_filer[-1, :]
             e = np.squeeze(p.e[-1, :, :])
         elif method == "TPI_scalar":
             tax_noncompliance = p.capital_income_tax_noncompliance_rate[0, :]
+            income_tax_filer = p.income_tax_filer[0, :]
+            wealth_tax_filer = p.wealth_tax_filer[0, :]
             e = np.squeeze(p.e[0, :, :])
         else:
             length = r.shape[0]
             tax_noncompliance = p.capital_income_tax_noncompliance_rate[
                 t : t + length, :
             ]
+            income_tax_filer = p.income_tax_filer[t : t + length, :]
+            wealth_tax_filer = p.wealth_tax_filer[t : t + length, :]
             e_long = np.concatenate(
                 (
                     p.e,
@@ -562,14 +616,17 @@ def FOC_savings(
         m_wealth = p.m_wealth[-1]
         p_wealth = p.p_wealth[-1]
         p_tilde = np.ones_like(p.rho[-1, :]) * p_tilde
+        tau_c = p.tau_c[-1, :]
     elif method == "TPI_scalar":
         h_wealth = p.h_wealth[0]
         m_wealth = p.m_wealth[0]
         p_wealth = p.p_wealth[0]
+        tau_c = p.tau_c[0, :]
     else:
-        h_wealth = p.h_wealth[t]
-        m_wealth = p.m_wealth[t]
-        p_wealth = p.p_wealth[t]
+        h_wealth = p.h_wealth[t : t + length]
+        m_wealth = p.m_wealth[t : t + length]
+        p_wealth = p.p_wealth[t : t + length]
+        tau_c = p.tau_c[t : t + length, :]
     taxes = tax.net_taxes(
         r,
         w,
@@ -588,7 +645,9 @@ def FOC_savings(
         etr_params,
         p,
     )
-    cons = get_cons(r, w, p_tilde, b, b_splus1, n, bq, rm, taxes, e, p)
+    cons = get_cons(
+        r, w, p_tilde, p_i, b, b_splus1, n, bq, rm, taxes, e, tau_c, p, method
+    )
     deriv = (
         (1 + r)
         - (
@@ -604,10 +663,11 @@ def FOC_savings(
                 etr_params,
                 mtry_params,
                 tax_noncompliance,
+                income_tax_filer,
                 p,
             )
         )
-        - tax.MTR_wealth(b, h_wealth, m_wealth, p_wealth)
+        - tax.MTR_wealth(b, h_wealth, m_wealth, p_wealth, wealth_tax_filer)
     )
     savings_ut = (
         rho * np.exp(-p.sigma * p.g_y) * chi_b * b_splus1 ** (-p.sigma)
@@ -641,6 +701,7 @@ def FOC_labor(
     r,
     w,
     p_tilde,
+    p_i,
     b,
     b_splus1,
     n,
@@ -675,6 +736,7 @@ def FOC_labor(
         r (array_like): the real interest rate
         w (array_like): the real wage rate
         p_tilde (array_like): composite good price
+        p_i (Numpy array): prices for consumption good i
         b (Numpy array): household savings
         b_splus1 (Numpy array): household savings one period ahead
         n (Numpy array): household labor supply
@@ -704,22 +766,28 @@ def FOC_labor(
     """
     if method == "SS":
         tau_payroll = p.tau_payroll[-1]
+        tau_c = p.tau_c[-1, :]
     elif method == "TPI_scalar":  # for 1st donut ring only
         tau_payroll = p.tau_payroll[0]
+        tau_c = p.tau_c[0, :]
     else:
         length = r.shape[0]
         tau_payroll = p.tau_payroll[t : t + length]
+        tau_c = p.tau_c[t : t + length, :]
     if j is not None:
         if method == "SS":
             tax_noncompliance = p.labor_income_tax_noncompliance_rate[-1, j]
+            income_tax_filer = p.income_tax_filer[-1, j]
             e = np.squeeze(p.e[-1, :, j])
         elif method == "TPI_scalar":
             tax_noncompliance = p.labor_income_tax_noncompliance_rate[0, j]
+            income_tax_filer = p.income_tax_filer[0, j]
             e = np.squeeze(p.e[0, -1, j])
         else:
             tax_noncompliance = p.labor_income_tax_noncompliance_rate[
                 t : t + length, j
             ]
+            income_tax_filer = p.income_tax_filer[t : t + length, j]
             e_long = np.concatenate(
                 (
                     p.e,
@@ -731,14 +799,17 @@ def FOC_labor(
     else:
         if method == "SS":
             tax_noncompliance = p.labor_income_tax_noncompliance_rate[-1, :]
+            income_tax_filer = p.income_tax_filer[-1, :]
             e = np.squeeze(p.e[-1, :, :])
         elif method == "TPI_scalar":
             tax_noncompliance = p.labor_income_tax_noncompliance_rate[0, :]
+            income_tax_filer = p.income_tax_filer[0, :]
             e = np.squeeze(p.e[0, -1, :])
         else:
             tax_noncompliance = p.labor_income_tax_noncompliance_rate[
                 t : t + length, :
             ]
+            income_tax_filer = p.income_tax_filer[t : t + length, :]
             e_long = np.concatenate(
                 (
                     p.e,
@@ -747,13 +818,6 @@ def FOC_labor(
                 axis=0,
             )
             e = np.diag(e_long[t : t + p.S, :, j], max(p.S - length, 0))
-    if method == "SS":
-        tau_payroll = p.tau_payroll[-1]
-    elif method == "TPI_scalar":  # for 1st donut ring only
-        tau_payroll = p.tau_payroll[0]
-    else:
-        length = r.shape[0]
-        tau_payroll = p.tau_payroll[t : t + length]
     if method == "TPI":
         if b.ndim == 2:
             r = r.reshape(r.shape[0], 1)
@@ -778,7 +842,9 @@ def FOC_labor(
         etr_params,
         p,
     )
-    cons = get_cons(r, w, p_tilde, b, b_splus1, n, bq, rm, taxes, e, p)
+    cons = get_cons(
+        r, w, p_tilde, p_i, b, b_splus1, n, bq, rm, taxes, e, tau_c, p, method
+    )
     deriv = (
         1
         - tau_payroll
@@ -793,6 +859,7 @@ def FOC_labor(
             etr_params,
             mtrx_params,
             tax_noncompliance,
+            income_tax_filer,
             p,
         )
     )